Progressive Enhancement
Javascript files and coding are used to allow users with Javascript
enabled in their browser to view an enhance version of the website not
accessible to users with Javascript disabled in their browser. Minor
changes have also been made to php file the output the result in the
form of HTML tags that are rendered to the browser.